🚚 중복 코드 kakao_api 파일로 이동 #11
